requirements.txt
Для бэкенда:

asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3

запуск локального сервера python3 manage.py runserver

Для фронтенда:

react
react-router-dom
axios

запуск локального сервера npm start

